Combined victory and death screen
Improved code in "core/_word.py" by commenting out debug code and making the amount of remaining tries more access able after the game is over.
Removed a bug in form of a spelling mistake in "core/word_handler.py" and added a function for getting information from the word after the game is over. For the death/victory page.
Commented debug code from "main.py" and combined the victory and death pages.
changed "ts/game.ts" to redirect the user to the game results page after the game is lost / won. And recompiled.